Move a service implementation into core to live with the service interface, as this service no longer represents a composite build specific view of the projects in the build tree.
