+++ subprojects/core/src/test/groovy/org/gradle/api/internal/artifacts/DefaultProjectComponentIdentifierTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013 the original author or authors.
+ * Copyright 2018 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,11 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.gradle.internal.component.local.model
+package org.gradle.api.internal.artifacts
 
 import org.gradle.api.artifacts.component.BuildIdentifier
 import org.gradle.api.artifacts.component.ProjectComponentIdentifier
-import org.gradle.api.internal.artifacts.DefaultBuildIdentifier
 import spock.lang.Specification
 import spock.lang.Unroll
